perf: use uint16_t for SegmentedLine coordinates instead of unsigned long
Screen coordinates never exceed 320x200, so 32-bit parameters were wasteful.
Shrinks each call from 4 long-pushes to 4 word-pushes (saves 8 bytes of stack
per call) and avoids the implicit +2 offset trick the assembly used to reach
the low word of each 32-bit slot.

+++ backend_gemtos.c
@@ -16,8 +16,8 @@
 
 /* from segline.s / clipline.s */
 void SegmentedLineSetup(void);
-void SegmentedLine(unsigned long x0, unsigned long y0,
-                   unsigned long x1, unsigned long y1, void *buffer);
+void SegmentedLine(uint16_t x0, uint16_t y0,
+                   uint16_t x1, uint16_t y1, void *buffer);
 void SegmentedMultiLine(Line *lines, void *buffer);
 
 static void  *gScreenBufferA;
